+++ src/NSchema.Core/Project/Ddl/DdlSyntaxException.cs
@@ -1,4 +1,4 @@
-using NSchema.Project.Ddl.Models;
+using NSchema.Project.Nsql;
 
 namespace NSchema.Project.Ddl;